Revert "chore: Move publisher snaps to application layout (#5036)"
This reverts commit f6ae120a210804c06a1f58fcda524df86b1ee92e.
