+++ static/js/publisher/components/PrimaryNav/PrimaryNav.tsx
@@ -55,7 +55,6 @@ function PrimaryNav({
                 label: "Overview",
                 href: "/snaps",
                 icon: "pods",
-                "aria-current": location.pathname === "/snaps",
               },
             ],
           },
