+++ static/js/publisher/publisher.tsx
@@ -13,7 +13,6 @@ import Listing from "./pages/Listing";
 import Builds from "./pages/Builds";
 import Build from "./pages/Build";
 import Releases from "./pages/Releases";
-import AccountSnaps from "./pages/AccountSnaps";
 
 const router = createBrowserRouter([
   {
@@ -63,10 +62,6 @@ const router = createBrowserRouter([
       {
         path: "/:snapId/releases",
         element: <Releases />,
-      },
-      {
-        path: "/snaps",
-        element: <AccountSnaps />,
       },
     ],
   },
