+++ static/js/snap-pages/components/EmptySnapList/EmptySnapList.tsx
