+++ static/js/snap-pages/components/NewSnapNotification/NewSnapNotification.tsx
