+++ static/js/snap-pages/components/NewSnapNotification/index.ts
