+++ static/js/snap-pages/components/PublishedSnapList/__tests__/PublishedSnapList.test.tsx
@@ -1,5 +1,3 @@
-import { BrowserRouter } from "react-router-dom";
-import { QueryClient, QueryClientProvider } from "react-query";
 import { render, screen } from "@testing-library/react";
 
 import PublishedSnapList from "../PublishedSnapList";
@@ -42,21 +40,15 @@ const BASE_SNAP_DATA = {
   unlisted: false,
 };
 
-const queryClient = new QueryClient();
-
 const renderComponent = (snaps: ISnap[]) => {
   return render(
-    <BrowserRouter>
-      <QueryClientProvider client={queryClient}>
-        <PublishedSnapList
-          currentPage={1}
-          currentUser="test-user-1"
-          setCurrentPage={() => {}}
-          snaps={snaps}
-          totalItems={snaps.length}
-        />
-      </QueryClientProvider>
-    </BrowserRouter>,
+    <PublishedSnapList
+      currentPage={1}
+      currentUser="test-user-1"
+      setCurrentPage={() => {}}
+      snaps={snaps}
+      totalItems={snaps.length}
+    />,
   );
 };
