+++ static/js/snap-pages/components/PublishedSnapSection/index.ts
