+++ static/js/snap-pages/components/RegisteredSnaps/RegisteredSnaps.tsx
@@ -1,5 +1,3 @@
-import { useState } from "react";
-import { Link } from "react-router-dom";
 import {
   ConfirmationModal,
   MainTable,
@@ -8,8 +6,8 @@ import {
   Strip,
   Tooltip,
 } from "@canonical/react-components";
-
 import { ISnap } from "../../types";
+import { useState } from "react";
 import { PAGE_NUMBER } from "../../types/constants";
 
 function RegisteredSnaps({
@@ -40,7 +38,7 @@ function RegisteredSnaps({
             width: "25%",
             content: (
               <>
-                <Link to={`/${snap.snapName}/listing`}>{snap.snapName}</Link>
+                <a href={`/${snap.snapName}/listing`}>{snap.snapName}</a>
                 {isDisputePending && (
                   <>
                     &nbsp;
@@ -90,9 +88,11 @@ function RegisteredSnaps({
                 (Name dispute in progress)
               </span>
             ) : (
-              <Link to="/docs/releasing-your-app" target="_blank">
-                Publish to this name
-              </Link>
+              <button className="p-button--link u-no-margin--bottom">
+                <a href="/docs/releasing-your-app" target="_blank">
+                  Publish to this name
+                </a>
+              </button>
             ),
             className: "u-align--right",
           },
