+++ static/js/snap-pages/components/RegisteredSnaps/index.ts
