+++ static/js/snap-pages/hooks/__tests__/useFetchAccountSnaps.test.ts
