+++ static/js/snap-pages/hooks/index.ts
@@ -1,0 +1,4 @@
+import useFetchAccountSnaps from "./useFetchAccountSnaps";
+import useFetchPublishedSnapMetrics from "./useFetchPublishedSnapMetrics";
+
+export { useFetchAccountSnaps, useFetchPublishedSnapMetrics };
