+++ static/js/snap-pages/pages/AccountSnaps/index.ts
