+++ static/js/snap-pages/routes/root.tsx
@@ -1,0 +1,16 @@
+import { Outlet } from "react-router-dom";
+import { Panel } from "@canonical/react-components";
+
+function Root(): JSX.Element {
+  return (
+    <>
+      <div id="main-content">
+        <Panel>
+          <Outlet />
+        </Panel>
+      </div>
+    </>
+  );
+}
+
+export default Root;
