+++ static/js/snap-pages/types/constants.ts
