+++ templates/publisher/account-snaps.html
@@ -1,0 +1,15 @@
+{% extends "publisher/_publisher_layout.html" %}
+
+{% block meta_title %}
+My published snaps — Linux software in the Snap Store
+{% endblock %}
+
+{% block content %}
+<div id="root">
+</div>
+
+<script src="{{ static_url('js/dist/snap-pages.js') }}"></script>
+<script>
+  window.CSRF_TOKEN = "{{ csrf_token() }}";
+</script>
+{% endblock %}
